refactor: 治理并精简 instrument test
