+++ .agents/skills/zhihu-instrument-test-governance/agents/openai.yaml
@@ -1,0 +1,4 @@
+interface:
+  display_name: "Zhihu++ Instrument Test Governance"
+  short_description: "Govern instrument tests by provenance and contract value"
+  default_prompt: "Audit Zhihu++ instrument tests, preserve every proven regression test, add verified issue and pull request provenance, and remove only low-value device coverage in a separate PR."
