+++ app/src/androidTest/java/com/github/zly2006/zhihu/AppearanceSettingsScreenInstrumentedTest.kt
@@ -19,20 +19,16 @@ package com.github.zly2006.zhihu
 
 import android.content.Context
 import android.content.SharedPreferences
-import androidx.compose.ui.semantics.SemanticsActions
 import androidx.compose.ui.test.SemanticsMatcher
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertTextContains
 import androidx.compose.ui.test.hasAnyDescendant
 import androidx.compose.ui.test.hasClickAction
-import androidx.compose.ui.test.hasTestTag
 import androidx.compose.ui.test.hasText
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onAllNodesWithTag
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
-import androidx.compose.ui.test.performScrollToNode
-import androidx.compose.ui.test.performSemanticsAction
 import androidx.compose.ui.test.performTouchInput
 import androidx.compose.ui.test.swipeUp
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -45,21 +41,13 @@ import com.github.zly2006.zhihu.navigation.OnlineHistory
 import com.github.zly2006.zhihu.test.performVerticalSwipeCycle
 import com.github.zly2006.zhihu.test.resetAppPreferences
 import com.github.zly2006.zhihu.test.setScreenContent
-import com.github.zly2006.zhihu.ui.ANSWER_DOUBLE_TAP_ACTION_PREFERENCE_KEY
-import com.github.zly2006.zhihu.ui.ARTICLE_USE_WEBVIEW_PREFERENCE_KEY
-import com.github.zly2006.zhihu.ui.AnswerDoubleTapAction
 import com.github.zly2006.zhihu.ui.PREFERENCE_NAME
-import com.github.zly2006.zhihu.ui.subscreens.APPEARANCE_SETTINGS_ANSWER_DOUBLE_TAP_TAG
 import com.github.zly2006.zhihu.ui.subscreens.APPEARANCE_SETTINGS_BOTTOM_BAR_SECTION_KEY
 import com.github.zly2006.zhihu.ui.subscreens.APPEARANCE_SETTINGS_COLLECTION_DIRECT_BROWSE_TAG
 import com.github.zly2006.zhihu.ui.subscreens.APPEARANCE_SETTINGS_SCROLL_TAG
-import com.github.zly2006.zhihu.ui.subscreens.APPEARANCE_SETTINGS_START_DESTINATION_TAG
-import com.github.zly2006.zhihu.ui.subscreens.APPEARANCE_SETTINGS_USE_WEBVIEW_TAG
 import com.github.zly2006.zhihu.ui.subscreens.AppearanceSettingsScreen
-import com.github.zly2006.zhihu.ui.subscreens.BOTTOM_BAR_ITEMS_PREFERENCE_KEY
 import com.github.zly2006.zhihu.ui.subscreens.BOTTOM_BAR_ITEM_ORDER_PREFERENCE_KEY
 import com.github.zly2006.zhihu.ui.subscreens.COLLECTION_DIRECT_BROWSE_PREFERENCE_KEY
-import com.github.zly2006.zhihu.ui.subscreens.START_DESTINATION_PREFERENCE_KEY
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
@@ -75,96 +63,10 @@ class AppearanceSettingsScreenInstrumentedTest {
     private val preferences: SharedPreferences
         get() = composeRule.activity.getSharedPreferences(PREFERENCE_NAME, Context.MODE_PRIVATE)
 
-    @Test
-    fun togglingWebViewModePersistsAndKeepsDependentControlsStableAfterScroll() {
-        // This test verifies a full user path in the answer section: scroll to the WebView setting,
-        // toggle it on, perform an extra swipe cycle, and prove that the persisted preference stays
-        // in sync before toggling it off again.
-        setUpScreen(setting = ARTICLE_USE_WEBVIEW_PREFERENCE_KEY)
-
-        waitUntilTagExists(APPEARANCE_SETTINGS_USE_WEBVIEW_TAG)
-        scrollUntilTagDisplayed(APPEARANCE_SETTINGS_USE_WEBVIEW_TAG)
-        assertFalse(preferences.getBoolean(ARTICLE_USE_WEBVIEW_PREFERENCE_KEY, false))
-        composeRule.onNodeWithTag(APPEARANCE_SETTINGS_USE_WEBVIEW_TAG).performClick()
-        waitUntilBooleanPreference(ARTICLE_USE_WEBVIEW_PREFERENCE_KEY, expected = true)
-
-        scrollContainer().performVerticalSwipeCycle()
-        waitUntilBooleanPreference(ARTICLE_USE_WEBVIEW_PREFERENCE_KEY, expected = true)
-
-        setUpScreen(
-            setting = ARTICLE_USE_WEBVIEW_PREFERENCE_KEY,
-            resetPreferences = false,
-        )
-        waitUntilTagExists(APPEARANCE_SETTINGS_USE_WEBVIEW_TAG)
-        scrollUntilTagDisplayed(APPEARANCE_SETTINGS_USE_WEBVIEW_TAG)
-        composeRule.onNodeWithTag(APPEARANCE_SETTINGS_USE_WEBVIEW_TAG).performClick()
-        waitUntilBooleanPreference(ARTICLE_USE_WEBVIEW_PREFERENCE_KEY, expected = false)
-    }
-
-    @Test
-    fun selectingAnswerDoubleTapActionUpdatesDropdownTextAndPreference() {
-        // This test verifies the dropdown click path for the answer double-tap action: after scrolling
-        // to the control, opening the anchored menu, and picking a different option, the visible label
-        // and the stored preference must both change to the same deterministic value.
-        setUpScreen(setting = ANSWER_DOUBLE_TAP_ACTION_PREFERENCE_KEY)
-
-        scrollUntilTagDisplayed(APPEARANCE_SETTINGS_ANSWER_DOUBLE_TAP_TAG)
-        composeRule
-            .onNodeWithTag(APPEARANCE_SETTINGS_ANSWER_DOUBLE_TAP_TAG)
-            .assertTextContains(AnswerDoubleTapAction.Ask.label)
-
-        composeRule.onNodeWithTag(APPEARANCE_SETTINGS_ANSWER_DOUBLE_TAP_TAG).performClick()
-        composeRule.onNode(hasText(AnswerDoubleTapAction.VoteUp.label), useUnmergedTree = true).performClick()
-
-        waitUntilStringPreference(
-            ANSWER_DOUBLE_TAP_ACTION_PREFERENCE_KEY,
-            expected = AnswerDoubleTapAction.VoteUp.preferenceValue,
-        )
-        composeRule
-            .onNodeWithTag(APPEARANCE_SETTINGS_ANSWER_DOUBLE_TAP_TAG)
-            .assertTextContains(AnswerDoubleTapAction.VoteUp.label)
-    }
-
-    @Test
-    fun bottomBarSelectionAndStartDestinationRemainStableAcrossScrollAndClicks() {
-        // This test verifies a multi-step bottom-bar configuration flow: remove one selected item,
-        // add another one, pick the new entry as the startup destination, and then perform an extra
-        // scroll cycle to ensure the rendered state still matches the persisted SharedPreferences.
-        setUpScreen(setting = APPEARANCE_SETTINGS_BOTTOM_BAR_SECTION_KEY)
-
-        scrollUntilTagDisplayed("appearanceSettings:bottomBar:item:${OnlineHistory.name}")
-        composeRule.onNodeWithTag("appearanceSettings:bottomBar:item:${OnlineHistory.name}").performClick()
-        waitUntilStringSetPreference(
-            BOTTOM_BAR_ITEMS_PREFERENCE_KEY,
-            expected = setOf(Home.name, Follow.name, Daily.name, Account.name),
-        )
-
-        composeRule.onNodeWithTag("appearanceSettings:bottomBar:item:${HotList.name}").performClick()
-        waitUntilStringSetPreference(
-            BOTTOM_BAR_ITEMS_PREFERENCE_KEY,
-            expected = setOf(Home.name, Follow.name, Daily.name, HotList.name, Account.name),
-        )
-
-        scrollContainer().performScrollToNode(hasTestTag(APPEARANCE_SETTINGS_START_DESTINATION_TAG))
-        composeRule
-            .onNodeWithTag(APPEARANCE_SETTINGS_START_DESTINATION_TAG)
-            .assertIsDisplayed()
-            .performSemanticsAction(SemanticsActions.OnClick)
-        val hotListOptionTag = "appearanceSettings:startDestination:option:${HotList.name}"
-        waitUntilTagDisplayed(hotListOptionTag)
-        composeRule.onNodeWithTag(hotListOptionTag).performClick()
-
-        waitUntilStringPreference(START_DESTINATION_PREFERENCE_KEY, expected = HotList.name)
-        scrollContainer().performVerticalSwipeCycle()
-        composeRule
-            .onNodeWithTag(APPEARANCE_SETTINGS_START_DESTINATION_TAG)
-            .assertTextContains("热榜")
-        assertEquals(
-            setOf(Home.name, Follow.name, Daily.name, HotList.name, Account.name),
-            preferences.getStringSet(BOTTOM_BAR_ITEMS_PREFERENCE_KEY, emptySet())?.toSet(),
-        )
-    }
-
+    /**
+     * Contract: https://github.com/zly2006/zhihu-plus-plus/issues/354
+     * Introduced by: https://github.com/zly2006/zhihu-plus-plus/pull/409
+     */
     @Test
     fun bottomBarRowsKeepUniformHeightAndMoveOrderPersists() {
         // The bottom-bar editor mixes selected rows, unselected rows, and the non-removable account
@@ -186,6 +88,10 @@ class AppearanceSettingsScreenInstrumentedTest {
         )
     }
 
+    /**
+     * Contract: https://github.com/zly2006/zhihu-plus-plus/issues/609
+     * Introduced by: https://github.com/zly2006/zhihu-plus-plus/pull/611
+     */
     @Test
     fun collectionDirectBrowseIsOptInAndClearlyMarkedAsExperimental() {
         setUpScreen(setting = COLLECTION_DIRECT_BROWSE_PREFERENCE_KEY)
