+++ app/src/androidTest/java/com/github/zly2006/zhihu/SettingsSearchScreenInstrumentedTest.kt
@@ -51,6 +51,10 @@ class SettingsSearchScreenInstrumentedTest {
         composeRule.resetAppPreferences()
     }
 
+    /**
+     * Contract: https://github.com/zly2006/zhihu-plus-plus/issues/262
+     * Introduced by: https://github.com/zly2006/zhihu-plus-plus/pull/455
+     */
     @Test
     fun searchScreen_filtersAppearanceResultsAndNavigatesToTargetSetting() {
         val navigator = setSearchScreenContent()
@@ -69,6 +73,10 @@ class SettingsSearchScreenInstrumentedTest {
         )
     }
 
+    /**
+     * Contract: https://github.com/zly2006/zhihu-plus-plus/issues/262
+     * Introduced by: https://github.com/zly2006/zhihu-plus-plus/pull/455
+     */
     @Test
     fun searchScreen_matchesPageNamesConfiguredByGroupedSettings() {
         val navigator = setSearchScreenContent()
@@ -87,6 +95,10 @@ class SettingsSearchScreenInstrumentedTest {
         )
     }
 
+    /**
+     * Contract: https://github.com/zly2006/zhihu-plus-plus/issues/262
+     * Introduced by: https://github.com/zly2006/zhihu-plus-plus/pull/455
+     */
     @Test
     fun searchScreen_canJumpToSystemAndNotificationSettingsEntries() {
         val navigator = setSearchScreenContent()
@@ -126,6 +138,10 @@ class SettingsSearchScreenInstrumentedTest {
         )
     }
 
+    /**
+     * Contract: https://github.com/zly2006/zhihu-plus-plus/issues/262
+     * Introduced by: https://github.com/zly2006/zhihu-plus-plus/pull/455
+     */
     @Test
     fun searchScreen_onlyExposesDeveloperEntryWhenDeveloperModeIsEnabled() {
         val hiddenNavigator = setSearchScreenContent()
